Remove effectively unnecessary `:nessie-jaxrs` module
