+++ gradle/projects.main.properties
@@ -17,7 +17,6 @@ nessie-gc-repository-jdbc=gc/gc-repository-jdbc
 nessie-gcs-testcontainer=testing/gcs-container
 nessie-perftest-gatling=perftest/gatling
 nessie-perftest-simulations=perftest/simulations
-nessie-jaxrs=servers/jax-rs
 nessie-jaxrs-testextension=servers/jax-rs-testextension
 nessie-jaxrs-tests=servers/jax-rs-tests
 nessie-keycloak-testcontainer=testing/keycloak-container
