--- servers/jax-rs/src/main/java/org/projectnessie/versioned/VersionStoreImplExtension.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (C) 2022 Dremio
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.projectnessie.versioned;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.enterprise.event.Observes;
-import jakarta.enterprise.inject.Default;
-import jakarta.enterprise.inject.spi.AfterBeanDiscovery;
-import jakarta.enterprise.inject.spi.BeanManager;
-import jakarta.enterprise.inject.spi.Extension;
-import jakarta.enterprise.util.TypeLiteral;
-import java.util.function.Supplier;
-import org.projectnessie.versioned.storage.common.persist.Persist;
-import org.projectnessie.versioned.storage.versionstore.VersionStoreImpl;
-
-public class VersionStoreImplExtension implements Extension {
-
-  private static Supplier<Persist> persist;
-
-  public static VersionStoreImplExtension forPersist(Supplier<Persist> persist) {
-    VersionStoreImplExtension.persist = persist;
-    return new VersionStoreImplExtension();
-  }
-
-  @SuppressWarnings("unused")
-  public void afterBeanDiscovery(@Observes AfterBeanDiscovery abd, BeanManager bm) {
-    abd.addBean()
-        .addType(new TypeLiteral<VersionStore>() {})
-        .addQualifier(Default.Literal.INSTANCE)
-        .scope(ApplicationScoped.class)
-        .produceWith(i -> new VersionStoreImpl(persist.get()));
-  }
-}
